nftigets_backend/requirements.txt
django
djangorestframework
django-cors-headers
django-environ
pytest
cx_Oracle
Pillow
xrpl-py
coverage
